Выносит валидацию в отдельный модуль
